Abstractify cameras; de-abstractify vectors; introduce antialiasing
Antialiasing is implemented by uniformly sampling each pixel. I am not
using random sampling because it is
a) one more thing to implement and
b) not entirely necessary yet

Compilations are becoming considerably slower; especially when the
number of rays cast increases by a factor of 16. I will eventually need
to find a bottleneck when compilation becomes far too slow.